requirements_other.txt
# PyTorch GPU + torchvision (latest with CUDA)

# CPU FAISS
faiss-cpu
# Core Python packages
absl-py>=1.0
argcomplete>=1.12.3
boto>=2.49.0
cached-property>=1.5.2
cachetools>=4.2.1
cffi>=1.14.6
chardet>=4.0.0
click>=8.0.1
crcmod>=1.7
cryptography>=39.0.0
cycler>=0.10.0
fasteners>=0.16.3
filelock>=3.0.12
gensim>=4.0.1
google-apitools>=0.5.32
google-auth>=2.20.0
google-auth-oauthlib>=1.0
google-reauth>=0.1.1
grpcio>=1.53.0
gymnasium
h5py>=3.8
higher>=0.2.1
httplib2>=0.19.1
huggingface-hub>=0.16
idna>=2.10
importlib-metadata>=6.0
joblib>=1.2.0
kiwisolver>=1.4.4
learn2learn
markdown>=3.3.3
matplotlib>=3.6
metann>=0.2.9
mock>=4.0.3
monotonic>=1.6
numpy>=1.24, <2.0.0
oauth2client>=4.1.3
oauthlib>=3.1.0
packaging>=21.3
pandas>=1.5
pbr>=5.11.1
progress>=1.5
protobuf>=3.20
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.20
pyglet>=2.0
pyopenssl>=23.0
pyparsing>=3.0
python-dateutil>=2.8.2
pyu2f>=0.1.5
pyyaml>=6.0
regex>=2021.11.10
requests>=2.28
requests-oauthlib>=1.3.1
retry-decorator>=1.1.1
rsa>=4.9
sacremoses>=0.0.53
scikit-learn>=1.2
scipy>=1.10
smart-open>=6.3.0
tensorboard>=2.12
tensorboard-plugin-wit>=1.8.1
tensorboardx>=2.6
threadpoolctl>=3.1
tokenizers>=0.13.3
tqdm>=4.65
transformers>=4.30
urllib3>=1.26
werkzeug>=2.3
zipp>=3.15
tensorboard
# Image processing
Pillow>=9.5

# Build / Compression
ninja>=1.11
lz4>=4.3

# Config / utility
easydict==1.9
termcolor==1.1.0